+++ src/utils.rs
@@ -1,6 +1,6 @@
 use std::{fmt::Display, time::Duration};
 
-pub fn result_message<T>(strategy: &str, iterations: u32, jobs: u32, time: Duration, piaprox: T)
+pub fn result_message<T>(strategy: &str, iterations: u64, jobs: u64, time: Duration, piaprox: T)
 where
     T: Display,
 {
